Created errors.py, move everything from main.py, moved Unit tests to sub directory, made talisman and app global for routes and errors
